% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.extract_from_se}
\alias{.extract_from_se}
\title{Extract miRNA and mRNA assays from SummarizedExperiment}
\usage{
.extract_from_se(se, mirna_assay = "mirna", mrna_assay = "mrna")
}
\arguments{
\item{se}{SummarizedExperiment instance}

\item{mirna_assay}{Character. miRNA assay name.}

\item{mrna_assay}{Character. mRNA assay name.}
}
\description{
Extract miRNA and mRNA assays from SummarizedExperiment
}
\keyword{internal}
