\name{fetch_mirBottleneck_data}
\alias{fetch_mirBottleneck_data}
\title{Fetch miR data}
\description{Fetches miR data based on specified parameters.}
\usage{fetch_mirBottleneck_data()}
\examples{
  \donttest{
    # Example of running the function
    data <- fetch_mirBottleneck_data()
    print("Data fetched successfully!")
  }
}