% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_pairwise_cor}
\alias{mean_pairwise_cor}
\title{Compute mean pairwise correlation of a matrix}
\usage{
mean_pairwise_cor(mat)
}
\arguments{
\item{mat}{Numeric matrix (genes x samples)}
}
\value{
Mean of upper triangle of correlation matrix, or NA
}
\description{
Compute mean pairwise correlation of a matrix
}
\keyword{internal}
