% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_mirna}
\alias{normalize_mirna}
\title{Normalize miRNA ID to lowercase precursor format}
\usage{
normalize_mirna(x)
}
\arguments{
\item{x}{Character vector of miRNA IDs}
}
\value{
Normalized character vector
}
\description{
Converts mature miRNA IDs (hsa-miR-21-5p) to lowercase precursor format
(hsa-mir-21) by lowercasing and stripping arm suffixes.
}
\examples{
normalize_mirna(c("hsa-miR-21-5p", "hsa-miR-21-3p"))
}
\keyword{internal}
