% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_bottleneck}
\alias{classify_bottleneck}
\title{Classify miRNAs and compute combined bottleneck index}
\usage{
classify_bottleneck(vss_scores, coherence_scores)
}
\arguments{
\item{vss_scores}{dataframe from score_vss()}

\item{coherence_scores}{dataframe from score_coherence()}
}
\value{
dataframe with mirna, vss, coherence_score, bottleneck_index, class
}
\description{
Classify miRNAs and compute combined bottleneck index
}
