% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_bottleneck}
\alias{classify_bottleneck}
\title{Classify miRNAs and compute combined bottleneck index}
\usage{
classify_bottleneck(vss_scores, coherence_scores)
}
\arguments{
\item{vss_scores}{dataframe from score_vss()}

\item{coherence_scores}{dataframe from score_coherence()}
}
\value{
dataframe with mirna, vss, coherence_score, bottleneck_index, class
}
\description{
Classify miRNAs and compute combined bottleneck index
}
\examples{
vss_scores <- data.frame(
  mirna = c("hsa-miR-21-5p", "hsa-miR-155-5p", "hsa-miR-1-3p"),
  vss = c(0.9, 0.4, 0.1)
)
coherence_scores <- data.frame(
  mirna = c("hsa-miR-21-5p", "hsa-miR-155-5p", "hsa-miR-1-3p"),
  coherence = c(0.8, 0.3, 0.2)
)
classify_bottleneck(vss_scores, coherence_scores)
}
